+++ etc/sample/app/src/org/jraf/androidcontentprovidergenerator/sample/provider/manual/ManualSelection.java
@@ -22,7 +22,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.jraf.androidcontentprovidergenerator.sample.provider.serialnumber;
+package org.jraf.androidcontentprovidergenerator.sample.provider.manual;
 
 import java.util.Date;
 
@@ -33,12 +33,12 @@ import android.net.Uri;
 import org.jraf.androidcontentprovidergenerator.sample.provider.base.AbstractSelection;
 
 /**
- * Selection for the {@code serial_number} table.
+ * Selection for the {@code manual} table.
  */
-public class SerialNumberSelection extends AbstractSelection<SerialNumberSelection> {
+public class ManualSelection extends AbstractSelection<ManualSelection> {
     @Override
     protected Uri baseUri() {
-        return SerialNumberColumns.CONTENT_URI;
+        return ManualColumns.CONTENT_URI;
     }
 
     /**
@@ -48,91 +48,91 @@ public class SerialNumberSelection extends AbstractSelection<SerialNumberSelecti
      * @param projection A list of which columns to return. Passing null will return all columns, which is inefficient.
      * @param sortOrder How to order the rows, formatted as an SQL ORDER BY clause (excluding the ORDER BY itself). Passing null will use the default sort
      *            order, which may be unordered.
-     * @return A {@code SerialNumberCursor} object, which is positioned before the first entry, or null.
+     * @return A {@code ManualCursor} object, which is positioned before the first entry, or null.
      */
-    public SerialNumberCursor query(ContentResolver contentResolver, String[] projection, String sortOrder) {
+    public ManualCursor query(ContentResolver contentResolver, String[] projection, String sortOrder) {
         Cursor cursor = contentResolver.query(uri(), projection, sel(), args(), sortOrder);
         if (cursor == null) return null;
-        return new SerialNumberCursor(cursor);
+        return new ManualCursor(cursor);
     }
 
     /**
      * Equivalent of calling {@code query(contentResolver, projection, null)}.
      */
-    public SerialNumberCursor query(ContentResolver contentResolver, String[] projection) {
+    public ManualCursor query(ContentResolver contentResolver, String[] projection) {
         return query(contentResolver, projection, null);
     }
 
     /**
      * Equivalent of calling {@code query(contentResolver, projection, null, null)}.
      */
-    public SerialNumberCursor query(ContentResolver contentResolver) {
+    public ManualCursor query(ContentResolver contentResolver) {
         return query(contentResolver, null, null);
     }
 
 
-    public SerialNumberSelection id(long... value) {
-        addEquals("serial_number." + SerialNumberColumns._ID, toObjectArray(value));
+    public ManualSelection id(long... value) {
+        addEquals("manual." + ManualColumns._ID, toObjectArray(value));
         return this;
     }
 
-    public SerialNumberSelection part0(String... value) {
-        addEquals(SerialNumberColumns.PART0, value);
+    public ManualSelection title(String... value) {
+        addEquals(ManualColumns.TITLE, value);
         return this;
     }
 
-    public SerialNumberSelection part0Not(String... value) {
-        addNotEquals(SerialNumberColumns.PART0, value);
+    public ManualSelection titleNot(String... value) {
+        addNotEquals(ManualColumns.TITLE, value);
         return this;
     }
 
-    public SerialNumberSelection part0Like(String... value) {
-        addLike(SerialNumberColumns.PART0, value);
+    public ManualSelection titleLike(String... value) {
+        addLike(ManualColumns.TITLE, value);
         return this;
     }
 
-    public SerialNumberSelection part0Contains(String... value) {
-        addContains(SerialNumberColumns.PART0, value);
+    public ManualSelection titleContains(String... value) {
+        addContains(ManualColumns.TITLE, value);
         return this;
     }
 
-    public SerialNumberSelection part0StartsWith(String... value) {
-        addStartsWith(SerialNumberColumns.PART0, value);
+    public ManualSelection titleStartsWith(String... value) {
+        addStartsWith(ManualColumns.TITLE, value);
         return this;
     }
 
-    public SerialNumberSelection part0EndsWith(String... value) {
-        addEndsWith(SerialNumberColumns.PART0, value);
+    public ManualSelection titleEndsWith(String... value) {
+        addEndsWith(ManualColumns.TITLE, value);
         return this;
     }
 
-    public SerialNumberSelection part1(String... value) {
-        addEquals(SerialNumberColumns.PART1, value);
+    public ManualSelection isbn(String... value) {
+        addEquals(ManualColumns.ISBN, value);
         return this;
     }
 
-    public SerialNumberSelection part1Not(String... value) {
-        addNotEquals(SerialNumberColumns.PART1, value);
+    public ManualSelection isbnNot(String... value) {
+        addNotEquals(ManualColumns.ISBN, value);
         return this;
     }
 
-    public SerialNumberSelection part1Like(String... value) {
-        addLike(SerialNumberColumns.PART1, value);
+    public ManualSelection isbnLike(String... value) {
+        addLike(ManualColumns.ISBN, value);
         return this;
     }
 
-    public SerialNumberSelection part1Contains(String... value) {
-        addContains(SerialNumberColumns.PART1, value);
+    public ManualSelection isbnContains(String... value) {
+        addContains(ManualColumns.ISBN, value);
         return this;
     }
 
-    public SerialNumberSelection part1StartsWith(String... value) {
-        addStartsWith(SerialNumberColumns.PART1, value);
+    public ManualSelection isbnStartsWith(String... value) {
+        addStartsWith(ManualColumns.ISBN, value);
         return this;
     }
 
-    public SerialNumberSelection part1EndsWith(String... value) {
-        addEndsWith(SerialNumberColumns.PART1, value);
+    public ManualSelection isbnEndsWith(String... value) {
+        addEndsWith(ManualColumns.ISBN, value);
         return this;
     }
 }